refactor(binding): move BSLLSBinding to a dedicated subpackage
BSLLSBinding is the public static service-locator / embedding facade
(used by the SonarQube plugin), not a bootstrap entry point. Living in
the root package it conflated the two roles and was the only reason the
root Application package depended on the diagnostics engine.

Move it (and its test) to `...languageserver.binding`, leaving the root
package as pure bootstrap (MainApplication -> cli only). The embedding
facade now has an explicit home whose engine access is justified by role.

BREAKING CHANGE: embedders must update the import to
com.github._1c_syntax.bsl.languageserver.binding.BSLLSBinding. The
SonarQube plugin reference must be updated in lockstep.
